refactor(tauri): split shell bootstrap modules
